small refactor, next is generalising the generate function
