+++ src/main.rs
@@ -1,16 +1,14 @@
 use clap::Parser;
-use colored::Colorize;
 
 mod config;
 mod install;
-mod generate_pacman;
-mod generate_yay;
+mod generate;
 
 #[derive(Parser)]
+#[command(version)]
 enum Frontier {
     Install(install::Install),
-    GeneratePacman(generate_pacman::GeneratePacman),
-    GenerateYay(generate_yay::GenerateYay),
+    Generate(generate::Generate),
 }
 
 fn main() {
@@ -20,15 +18,6 @@ fn main() {
     // perform specific subcommand
     match args {
         Frontier::Install(install) => install.run(),
-        Frontier::GeneratePacman(generate_pacman) => {
-                println!("{} creating output file...", "[frontier]".bold().purple());
-                generate_pacman.run();
-                println!("{} success!", "[frontier]".bold().purple());
-            },
-        Frontier::GenerateYay(generate_yay) => {
-                println!("{} creating output file...", "[frontier]".bold().purple());
-                generate_yay.run();
-                println!("{} success!", "[frontier]".bold().purple());
-            },
+        Frontier::Generate(generate) => generate.run(),
     }
 }
